Page Flip App supports arbitrary Page components now
