+++ src/components/Book.jsx
@@ -1,8 +1,6 @@
 import styled from "@emotion/styled";
 import PageFlipApp from "./book-components/PageFlipApp";
 
-import TransitionComponent from "./book-components/TransitionComponent";
-
 const Book = ({ data, mobile }) => {
   const Container = styled.div`
     background-color: lightblue;
@@ -11,7 +9,7 @@ const Book = ({ data, mobile }) => {
 
   return (
     <Container>
-      <TransitionComponent />
+      <PageFlipApp />
     </Container>
   );
 };
